ipsecrps [0/1]: Add configuration cell for IPSec RPS mode
This commit introduces a new configuratin cell for a newly added IPSec
feature: Received Packet Steering. The cell added in this commit allows
for users to enable the feature and performs configuration validation,
ensuring that when the feature is enabled, IPSec and XDP Native mode are
also enabled.

The next commit contains the datapath implementation.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>